Move serializers to common_serializers, fix mypy
